requirements.dev.txt
# dev dependencies
ipython>=7.0.0,<=8.0.0
ipdb
pudb
pytest-watch

# code quality
flake8
pyproject-flake8
black
isort

# install project as editable
-e .